refactor(clawdstrike): extract evaluate decision core from engine.rs
Move the core decision flow into evaluate.rs as the final impl HushEngine
block: prepare_origin_context, check_action_report_prepared (the
BuiltIn -> Custom -> Extra -> Async pipeline + verdict aggregation),
evaluate_guard_stage, observe_guard_result, single_result_report, and the
report-metadata helpers. The guard-pipeline ordering invariant is moved
line-for-line.

mod.rs is now just the HushEngine/EngineState/PreparedContext/
PreparedEvaluation type defs, the trivial accessors, module decls and the
re-export surface (~112 lines). Test-only guard/origin types moved to
tests/mod.rs shared imports. Public API unchanged.

Pure relocation: 825 lib tests + 44 formal-diff-tests pass; clippy clean.
This completes the engine.rs god-file decomposition (step 6).
